# .agent/tmp/robyn_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OAP LangGraph Runtime",
    "version": "0.1.0",
    "description": "\nRobyn-based LangGraph-compatible runtime for Open Agent Platform.\n\nThis API provides endpoints for managing assistants, threads, runs, and the store.\nIt implements the LangGraph API specification for compatibility with LangGraph clients.\n"
  },
...
  "paths": {
...
    "/threads": {
      "post": {
        "tags": ["Threads"],
        "summary": "Create Thread",
        "description": "Create a new thread for running conversations.",
        "operationId": "create_thread",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreadCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thread created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Thread" }
              }
            }
          },
          "409": {
            "description": "Conflict - Thread already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/search": {
      "post": {
        "tags": ["Threads"],
        "summary": "Search Threads",
        "description": "Search for threads matching the specified criteria.",
        "operationId": "search_threads",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreadSearchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of matching threads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Thread" }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/count": {
      "post": {
        "tags": ["Threads"],
        "summary": "Count Threads",
        "description": "Count threads matching the specified criteria.",
        "operationId": "count_threads",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreadCountRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Count of matching threads",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "title": "Count" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}": {
      "get": {
        "tags": ["Threads"],
        "summary": "Get Thread",
        "description": "Retrieve a thread by its ID.",
        "operationId": "get_thread",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested thread",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Thread" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Threads"],
        "summary": "Update Thread",
        "description": "Update an existing thread's metadata.",
        "operationId": "update_thread",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreadPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated thread",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Thread" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Threads"],
        "summary": "Delete Thread",
        "description": "Delete a thread by its ID.",
        "operationId": "delete_thread",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          }
        ],
        "responses": {
          "200": {
            "description": "Thread deleted successfully",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}/state": {
      "get": {
        "tags": ["Threads"],
        "summary": "Get Thread State",
        "description": "Get the current state of a thread.",
        "operationId": "get_thread_state",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          }
        ],
        "responses": {
          "200": {
            "description": "The current thread state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ThreadState" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/threads/{thread_id}/history": {
      "get": {
        "tags": ["Threads"],
        "summary": "Get Thread History",
        "description": "Get the state history of a thread.",
        "operationId": "get_thread_history",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the thread."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 },
            "description": "Maximum number of history entries to return."
          }
        ],
        "responses": {
          "200": {
            "description": "List of thread states",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ThreadState" }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        },
        "title": "ErrorResponse",
        "description": "Standard error response format."
      },
...
      "Thread": {
        "type": "object",
        "required": [
          "thread_id",
          "created_at",
          "updated_at",
          "metadata",
          "status"
        ],
        "properties": {
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread ID",
            "description": "The unique identifier of the thread."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The time the thread was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The last time the thread was updated."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Custom metadata for the thread."
          },
          "config": {
            "type": "object",
            "title": "Config",
            "description": "Thread configuration."
          },
          "status": {
            "type": "string",
            "enum": ["idle", "busy", "interrupted", "error"],
            "title": "Status",
            "description": "The current status of the thread."
          },
          "values": {
            "type": "object",
            "title": "Values",
            "description": "The current state values of the thread."
          },
          "interrupts": {
            "type": "object",
            "title": "Interrupts",
            "description": "Active interrupts on the thread."
          }
        },
        "title": "Thread"
      },
      "ThreadCreate": {
        "type": "object",
        "properties": {
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread ID",
            "description": "The ID of the thread. If not provided, a random UUID will be generated."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Metadata to add to thread."
          },
          "if_exists": {
            "type": "string",
            "enum": ["raise", "do_nothing"],
            "title": "If Exists",
            "description": "How to handle duplicate creation.",
            "default": "raise"
          }
        },
        "title": "ThreadCreate",
        "description": "Payload for creating a thread."
      },
      "ThreadPatch": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Metadata to merge with existing thread metadata."
          }
        },
        "title": "ThreadPatch",
        "description": "Payload for updating a thread."
      },
      "ThreadSearchRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "title": "IDs",
            "description": "Filter by specific thread IDs."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Filter by metadata key-value pairs."
          },
          "values": {
            "type": "object",
            "title": "Values",
            "description": "Filter by state values."
          },
          "status": {
            "type": "string",
            "enum": ["idle", "busy", "interrupted", "error"],
            "title": "Status",
            "description": "Filter by thread status."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum number of results to return.",
            "default": 10,
            "minimum": 1,
            "maximum": 1000
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Number of results to skip.",
            "default": 0,
            "minimum": 0
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "description": "Field to sort by."
          },
          "sort_order": {
            "type": "string",
            "enum": ["asc", "desc"],
            "title": "Sort Order",
            "description": "Sort order."
          }
        },
        "title": "ThreadSearchRequest",
        "description": "Request body for searching threads."
      },
      "ThreadCountRequest": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Filter by metadata key-value pairs."
          },
          "values": {
            "type": "object",
            "title": "Values",
            "description": "Filter by state values."
          },
          "status": {
            "type": "string",
            "enum": ["idle", "busy", "interrupted", "error"],
            "title": "Status",
            "description": "Filter by thread status."
          }
        },
        "title": "ThreadCountRequest",
        "description": "Request body for counting threads."
      },
      "ThreadState": {
        "type": "object",
        "required": ["values", "next", "tasks"],
        "properties": {
          "values": {
            "anyOf": [
              { "type": "object" },
              { "type": "array", "items": { "type": "object" } }
            ],
            "title": "Values",
            "description": "The current state values."
          },
          "next": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Next",
            "description": "The next nodes to execute."
          },
          "tasks": {
            "type": "array",
            "items": { "type": "object" },
            "title": "Tasks",
            "description": "Pending tasks."
          },
          "checkpoint": {
            "type": "object",
            "title": "Checkpoint",
            "description": "The current checkpoint."
          },
          "metadata": { "type": "object", "title": "Metadata" },
          "created_at": { "type": "string", "title": "Created At" },
          "parent_checkpoint": {
            "type": "object",
            "title": "Parent Checkpoint"
          },
          "interrupts": {
            "type": "array",
            "items": { "type": "object" },
            "title": "Interrupts"
          }
        },
        "title": "ThreadState"
      },
...
    }
  }
}